[Fix] actionsを整理する
