+++ src/actions/random-cat.ts
@@ -1,0 +1,28 @@
+import { selectRandomCat, catSearchableText, randomCatRequested } from '../Cat';
+
+import { fetchAnimalImageUrl } from '../flicker/api';
+import { postImage } from '../slack/api';
+
+/*
+ * ランダムにニャンコを抽出します
+ */
+export async function randomCat(text: string): Promise<void> {
+  if (!randomCatRequested(text)) {
+    return;
+  }
+
+  const cat = selectRandomCat();
+  const searchableText = catSearchableText(cat)
+  const imageUrl = await fetchAnimalImageUrl(searchableText);
+
+  if (imageUrl === null) {
+    console.log('no images found!');
+    return;
+  }
+
+  try {
+    await postImage(imageUrl, cat);
+  } catch (e) {
+    console.log(`error occurred: ${e}`);
+  }
+}
